- tweak: Knife tweaks.
