+++ Content.Shared/_White/BetrayalDagger/BlinkComponent.cs
@@ -1,4 +1,6 @@
+using System.Numerics;
 using Robust.Shared.Audio;
+using Robust.Shared.Serialization;
 
 namespace Content.Shared._White.BetrayalDagger;
 
@@ -18,4 +20,17 @@ public sealed partial class BlinkComponent : Component
     {
         Params = AudioParams.Default.WithVolume(5f)
     };
+}
+
+[Serializable, NetSerializable]
+public sealed class BlinkEvent : EntityEventArgs
+{
+    public readonly NetEntity Weapon;
+    public readonly Vector2 Direction;
+
+    public BlinkEvent(NetEntity weapon, Vector2 direction)
+    {
+        Weapon = weapon;
+        Direction = direction;
+    }
 }
