+++ Resources/Prototypes/Entities/Objects/Weapons/Melee/daggers.yml
@@ -31,7 +31,7 @@
 - type: entity
   name: предательский нож
   description: Берегите спину.
-  parent: BaseItem
+  parent: BaseKnife
   id: BetrayalKnife
   components:
     - type: Sharp
@@ -40,14 +40,24 @@
       state: icon
     - type: MeleeWeapon
       wideAnimationRotation: 180
-      attackRate: 1
+      attackRate: 1.5
       damage:
         types:
-          Slash: 30
+          Slash: 17.5
       soundHit:
         path: /Audio/Weapons/bladeslice.ogg
+    - type: EmbeddableProjectile
+      sound: /Audio/Weapons/star_hit.ogg
+    - type: ThrowingAngle
+      angle: 180
+    - type: DamageOtherOnHit
+      damage:
+        types:
+          Slash: 20
     - type: Item
       size: Small
     - type: DisarmMalus
+      malus: 0.225
     - type: Backstab
     - type: Blink
+      blinkRate: 0.33
